requirements.txt
# Core Testing Framework
pytest==8.3.2
pytest-xdist==3.6.1
pytest-html==4.1.1
allure-pytest==2.13.5

# Selenium & WebDriver
selenium==4.25.0
webdriver-manager==4.0.2

# Config and Utilities
PyYAML==6.0.2
Faker==30.3.0
requests==2.32.3
python-dotenv==1.0.1

# Optional Helpers (Logging, Reports)
colorlog==6.8.2
